Add a long cast for the curl ip resolve options in sendmail code
This is correct per documentation and fixes a warning when compiling with
optimisation.

+++ src/sendmail.c
@@ -919,9 +919,9 @@ static CURL *newSendmailHandle(const struct MACCOUNT *account,
 	if(curlCiphers)
 		curl_easy_setopt(handle, CURLOPT_SSL_CIPHER_LIST, curlCiphers);
 	if(curlIPV == 4)
-		curl_easy_setopt(handle, CURLOPT_IPRESOLVE, CURL_IPRESOLVE_V4);
+		curl_easy_setopt(handle, CURLOPT_IPRESOLVE, (long)CURL_IPRESOLVE_V4);
 	if(curlIPV == 6)
-		curl_easy_setopt(handle, CURLOPT_IPRESOLVE, CURL_IPRESOLVE_V6);
+		curl_easy_setopt(handle, CURLOPT_IPRESOLVE, (long)CURL_IPRESOLVE_V6);
 	curl_easy_setopt(handle, CURLOPT_ERRORBUFFER, cerror);
 	curl_easy_setopt(handle, CURLOPT_CONNECTTIMEOUT, mailTimeout);
 	res = setCurlURL(handle, outurl);
